# internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for managing HydraDNS server configuration and filtering.",
        "title": "HydraDNS Management API",
        "contact": {
            "name": "HydraDNS Support",
            "url": "https://github.com/jroosing/hydradns"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/config": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns the current server configuration (sensitive fields redacted)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "Get current configuration",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ConfigResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates server configuration (requires restart for some settings)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "Update configuration",
                "parameters": [
                    {
                        "description": "Configuration update",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ConfigResponse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.StatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/config/reload": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Triggers a hot reload of configuration from disk",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "Reload configuration",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.StatusResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_jroosing_hydradns_internal_api_models.APIConfigResponse": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "host": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_jroosing_hydradns_internal_api_models.ConfigResponse": {
            "type": "object",
            "properties": {
                "api": {
                    "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.APIConfigResponse"
                },
                "custom_dns": {
                    "$ref": "#/definitions/github_com_jroosing_hydradns_internal_config.CustomDNSConfig"
                },
                "filtering": {
                    "$ref": "#/definitions/github_com_jroosing_hydradns_internal_config.FilteringConfig"
                },
                "logging": {
                    "$ref": "#/definitions/github_com_jroosing_hydradns_internal_config.LoggingConfig"
                },
                "rate_limit": {
                    "$ref": "#/definitions/github_com_jroosing_hydradns_internal_config.RateLimitConfig"
                },
                "server": {
                    "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ServerConfigResponse"
                },
                "upstream": {
                    "$ref": "#/definitions/github_com_jroosing_hydradns_internal_config.UpstreamConfig"
                }
            }
        },
...
        "github_com_jroosing_hydradns_internal_api_models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "github_com_jroosing_hydradns_internal_api_models.ServerConfigResponse": {
            "type": "object",
            "properties": {
                "enable_tcp": {
                    "type": "boolean"
                },
                "host": {
                    "type": "string"
                },
                "max_concurrency": {
                    "type": "integer"
                },
                "port": {
                    "type": "integer"
                },
                "tcp_fallback": {
                    "type": "boolean"
                },
                "upstream_socket_pool_size": {
                    "type": "integer"
                },
                "workers": {
                    "type": "string"
                }
            }
        },
...
        "github_com_jroosing_hydradns_internal_api_models.StatusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
...
        "github_com_jroosing_hydradns_internal_config.BlocklistConfig": {
            "type": "object",
            "properties": {
                "format": {
                    "description": "\"auto\", \"adblock\", \"hosts\", \"domains\"",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "github_com_jroosing_hydradns_internal_config.CustomDNSConfig": {
            "type": "object",
            "properties": {
                "cnames": {
                    "description": "CNAMEs maps alias names to canonical names\nExample: \"www.homelab.local\": \"homelab.local\"",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "hosts": {
                    "description": "Hosts maps domain names to IP addresses (IPv4 or IPv6)\nExample: \"homelab.local\": \"192.168.1.10\" or \"server.local\": \"2001:db8::1\"",
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "github_com_jroosing_hydradns_internal_config.FilteringConfig": {
            "type": "object",
            "properties": {
                "blacklist_domains": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "blocklists": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_jroosing_hydradns_internal_config.BlocklistConfig"
                    }
                },
                "enabled": {
                    "type": "boolean"
                },
                "log_allowed": {
                    "type": "boolean"
                },
                "log_blocked": {
                    "type": "boolean"
                },
                "refresh_interval": {
                    "type": "string"
                },
                "whitelist_domains": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_jroosing_hydradns_internal_config.LoggingConfig": {
            "type": "object",
            "properties": {
                "extra_fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "include_pid": {
                    "type": "boolean"
                },
                "level": {
                    "type": "string"
                },
                "structured": {
                    "type": "boolean"
                },
                "structured_format": {
                    "type": "string"
                }
            }
        },
        "github_com_jroosing_hydradns_internal_config.RateLimitConfig": {
            "type": "object",
            "properties": {
                "cleanup_seconds": {
                    "description": "CleanupSeconds is how often stale entries are cleaned up (default: 60)",
                    "type": "number"
                },
                "global_burst": {
                    "description": "GlobalBurst is the global burst size (default: 100000)",
                    "type": "integer"
                },
                "global_qps": {
                    "description": "GlobalQPS is the server-wide queries per second limit (default: 100000, 0 = disabled)",
                    "type": "number"
                },
                "ip_burst": {
                    "description": "IPBurst is the per-IP burst size (default: 6000)",
                    "type": "integer"
                },
                "ip_qps": {
                    "description": "IPQPS is the per-IP QPS limit (default: 3000, 0 = disabled)",
                    "type": "number"
                },
                "max_ip_entries": {
                    "description": "MaxIPEntries is the maximum number of tracked IPs (default: 65536)",
                    "type": "integer"
                },
                "max_prefix_entries": {
                    "description": "MaxPrefixEntries is the maximum number of tracked prefixes (default: 16384)",
                    "type": "integer"
                },
                "prefix_burst": {
                    "description": "PrefixBurst is the per-prefix burst size (default: 20000)",
                    "type": "integer"
                },
                "prefix_qps": {
                    "description": "PrefixQPS is the per-prefix QPS limit (default: 10000, 0 = disabled)",
                    "type": "number"
                }
            }
        },
        "github_com_jroosing_hydradns_internal_config.UpstreamConfig": {
            "type": "object",
            "properties": {
                "max_retries": {
                    "description": "Max retries per upstream on timeout",
                    "type": "integer"
                },
                "servers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tcp_timeout": {
                    "description": "Timeout for TCP queries (e.g., \"5s\")",
                    "type": "string"
                },
                "udp_timeout": {
                    "description": "Timeout for UDP queries (e.g., \"3s\")",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}